+++ src/snowflake/cli/api/project/schemas/entities/entities.py
@@ -38,6 +38,12 @@ from snowflake.cli._plugins.spcs.compute_pool.compute_pool_entity import (
 from snowflake.cli._plugins.spcs.compute_pool.compute_pool_entity_model import (
     ComputePoolEntityModel,
 )
+from snowflake.cli._plugins.spcs.image_repository.image_repository_entity import (
+    ImageRepositoryEntity,
+)
+from snowflake.cli._plugins.spcs.image_repository.image_repository_entity_model import (
+    ImageRepositoryEntityModel,
+)
 from snowflake.cli._plugins.streamlit.streamlit_entity import StreamlitEntity
 from snowflake.cli._plugins.streamlit.streamlit_entity_model import (
     StreamlitEntityModel,
@@ -50,6 +56,7 @@ Entity = Union[
     ProcedureEntity,
     FunctionEntity,
     ComputePoolEntity,
+    ImageRepositoryEntity,
 ]
 EntityModel = Union[
     ApplicationEntityModel,
@@ -58,6 +65,7 @@ EntityModel = Union[
     FunctionEntityModel,
     ProcedureEntityModel,
     ComputePoolEntityModel,
+    ImageRepositoryEntityModel,
 ]
 
 ALL_ENTITIES: List[Entity] = [*get_args(Entity)]